Refs #250 HL7 support
I see I already have a HL7 interface running on OpenREM for more than 7 years. It really is time to make this publicly available.

This update applies all changes needed and is rebased to the current development-branch. It is not yet tested (hopefully I just copied everything correct and then it should work).